+++ src/client/java/com/github/theredbrain/inventorysizeattributes/InventorySizeAttributesClient.java
@@ -1,10 +1,15 @@
 package com.github.theredbrain.inventorysizeattributes;
 
+import com.github.theredbrain.inventorysizeattributes.config.ClientConfig;
+import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
+import me.fzzyhmstrs.fzzy_config.api.RegisterType;
 import net.fabricmc.api.ClientModInitializer;
 
 public class InventorySizeAttributesClient implements ClientModInitializer {
+	public static ClientConfig CLIENT_CONFIG;
 
 	@Override
 	public void onInitializeClient() {
+		CLIENT_CONFIG = ConfigApiJava.registerAndLoadConfig(ClientConfig::new, RegisterType.CLIENT);
 	}
 }