+++ src/client/java/com/github/theredbrain/inventorysizeattributes/gui/screen/ingame/HasBackgroundWithNoSlots.java
@@ -1,0 +1,7 @@
+package com.github.theredbrain.inventorysizeattributes.gui.screen.ingame;
+
+import net.minecraft.util.Identifier;
+
+public interface HasBackgroundWithNoSlots {
+	Identifier inventorysizeattributes$getBackgroundWithNoSlots();
+}