+++ src/main/java/com/github/theredbrain/inventorysizeattributes/config/ClientConfig.java
@@ -1,0 +1,15 @@
+package com.github.theredbrain.inventorysizeattributes.config;
+
+import com.github.theredbrain.inventorysizeattributes.InventorySizeAttributes;
+import me.fzzyhmstrs.fzzy_config.config.Config;
+import me.fzzyhmstrs.fzzy_config.validation.misc.ValidatedBoolean;
+
+public class ClientConfig extends Config {
+
+	public ClientConfig() {
+		super(InventorySizeAttributes.identifier("client"));
+	}
+
+	public ValidatedBoolean show_inactive_inventory_slots = new ValidatedBoolean(true);
+
+}